requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
aiofiles==23.1.0

# ASR (choose one or both)
openai==1.11.0
whisper==20230314
faster-whisper==0.6.1

# Translation
transformers==4.35.0
sentencepiece==0.1.99
torch==2.2.0

# TTS (Coqui)
TTS==0.13.1

# Frontend
streamlit==1.37.1
requests==2.31.0

# Helpers
pydub==0.25.1
numpy==1.26.0
python-dotenv==1.0.1
